template/requirements.txt
-i https://pypi.doubanio.com/simple
ruamel.yaml==0.16.13
aiohttp==3.7.4
aiohttp-cors==0.7.0
# mxnet==1.6.0
# numpy==1.17.3

# numpy<1.20.0,>=1.14.0
cnocr==1.2.2   
paho-mqtt==1.5.1
aiomysql==0.0.21
asyncpg==0.22.0
# numpy >= 1.17.3
# opencv-python==4.5.1.48

# numpy >= 1.17.3
# pyanime4k==2.5.2


aioredis==1.3.1
# motor.motor_asyncio.AsyncIOMotorClient() 连接mongo
motor==2.3.1


gunicorn==20.0.4
